refactor(compliance): consolidate contributor check CLIs
